:hammer: refactor: MatchTestUtils util 폴더로 이동
